Update core.lua
* Updated UNIT_DIED event to check if Party/Raid Hunter(s) cast Feign Death or in fact died.

The only reliable method to use is Blizzards API UnitIsFeignDeath() which requires a UnitID and since the Combat Log API's only return GUID's this is impossible unless the Hunter being checked is in a Party or Raid with an addon user. If the Hunter is FD it returns true and bypasses the buffCache clears keeping the timers intact. If the Hunter dies then it returns false and clears the buffCache/timers as normal.